rename stuff in qa-tools
