+++ pallets/qa-tools/src/pallet_tools/mod.rs
@@ -1,4 +1,4 @@
 //! Pallet-separated tools. To avoid naming clashes and
 //! to distinguish logic related to different pallets.
 
-pub mod order_book_tools;
+pub mod order_book;
